+++ src/components/CustomDrawerNavigator/index.ts
@@ -1,4 +1,0 @@
-import CustomDrawerNavigator from './CustomDrawerNavigator';
-
-
-export default CustomDrawerNavigator;