+++ src/redux/reducers/ToggleDarkTheme.ts
@@ -1,14 +1,8 @@
 import { isType } from "typescript-fsa";
 
-import { THEMES } from "~/constants";
 import { ITheme } from "@types";
+import { THEMES } from "~/constants";
 import { toggleDarkTheme } from "~/redux/actions";
-
-/**
- * This reducer toggles whether to use the dark theme or not.
- * If set to true uses the dark theme across the app. Else
- * uses light theme. This reducer takes in true or false.
- */
 
 const initialState = {
   theme: THEMES.LIGHT_THEME