+++ src/services/web/middleware/index.ts
@@ -1,0 +1,3 @@
+import ValidateToken from "./Token";
+
+export { ValidateToken };