+++ CHANGELOG.md
@@ -6,3 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Added
+- Three encoding algorithms LSB, JSTEG & F5.
+- Steganography algorithms as FAAS on firebase.
+- Firebase integration into the application.
+- Initial React Native application version 1.0.0.